fixed bug in disassembler; reduced VC++ compiler warnings
git-svn-id: https://svn.code.sf.net/p/gk3tools/code/trunk@142 80473309-fabf-408f-887a-d98baf40245d

+++ sheep/sheepCodeGenerator.h
@@ -61,8 +61,8 @@ private:
 	std::vector<SheepImport> m_usedImports;
 	std::vector<SheepSymbol> m_variables;
 	
-	size_t getIndexOfImport(SheepImport& import);
-	size_t getIndexOfVariable(SheepSymbol& symbol);
+	int getIndexOfImport(SheepImport& import);
+	int getIndexOfVariable(SheepSymbol& symbol);
 
 	typedef std::map<std::string, SheepSymbol> SymbolMap;
 	SymbolMap m_symbolMap;
